# hackathonApp/backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Hackathon App API
  description: |
    REST API for the Hackathon Application. Supports user authentication (with optional
    MFA), ticket management, real-time chat (channels, DMs, ticket threads), Workday
    ad-hoc sync for admin users, and MFA administration.

    ## Authentication
    All authenticated endpoints require a valid session cookie (`hackathon_session`).
    Obtain a session by calling `POST /api/auth/login`. If the user has MFA enabled,
    a two-step login flow is required (see MFA section).

    ## Rate Limiting
    All endpoints are rate-limited. Unauthenticated endpoints are limited per IP;
    authenticated endpoints are limited per user. When rate-limited, the API returns
    `429 Too Many Requests` with a `Retry-After` header.

    ## Error Format
    All errors follow a consistent JSON format:
    ```json
    {
      "error": "error_code_or_message",
      "message": "Human-readable description"
    }
    ```
  version: 1.0.0
  contact:
    name: Hackathon App Team
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://hackathon.example.com/api
    description: Production server
...
paths:
...
  /tickets:
    get:
      tags: [Tickets]
      summary: List tickets
      security:
        - sessionCookie: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, resolved, closed]
      responses:
        '200':
          description: Ticket list
    post:
      tags: [Tickets]
      summary: Create a new ticket
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description]
              properties:
                description:
                  type: string
                  maxLength: 5000
                location_table:
                  type: string
                priority:
                  type: string
                  enum: [low, normal, high, urgent]
                  default: normal
      responses:
        '201':
          description: Ticket created
        '429':
          description: Rate limited (20 per hour per user)

  # ──────────────────────────────────────────────
  # WORKERS
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: hackathon_session
...